feat(#139) : regular_checker 폴더명 checker_regular로 변경함에 따른 경로 수정
