+++ everyday/src/main/java/com/nyam/everyday/module/challenge/checker/checker_regular/login/Login30DaysChecker.java
@@ -1,4 +1,4 @@
-package com.nyam.everyday.module.challenge.checker.regular_checker.login;
+package com.nyam.everyday.module.challenge.checker.checker_regular.login;
 
 import com.nyam.everyday.module.challenge.checker.AbstractDateBasedChecker;
 import com.nyam.everyday.module.challenge.checker.service.ChallengeCheckService;
@@ -8,28 +8,26 @@ import com.nyam.everyday.module.challenge.entity.ChallengeTag;
 import com.nyam.everyday.module.challenge.repository.ChallengeRepository;
 import com.nyam.everyday.module.challenge.repository.MemberChallengeDayRepository;
 import com.nyam.everyday.module.member.entity.Member;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
 
-@Slf4j
 @Component
-public class LoginFirstChecker extends AbstractDateBasedChecker {
+public class Login30DaysChecker extends AbstractDateBasedChecker {
 
-  protected LoginFirstChecker(
-          ChallengeRepository challengeRepository,
-          ChallengeCheckService challengeCheckService,
+  protected Login30DaysChecker(
           MemberChallengeDayRepository memberChallengeDayRepository,
-          ApplicationEventPublisher publisher
+          ApplicationEventPublisher publisher,
+          ChallengeRepository challengeRepository,
+          ChallengeCheckService challengeCheckService
   ) {
     super(challengeRepository, challengeCheckService, memberChallengeDayRepository, publisher);
   }
 
   @Override
   public ChallengeCode getChallengeCode() {
-    return ChallengeCode.LOGIN_FIRST;
+    return ChallengeCode.LOGIN_30DAYS;
   }
 
   @Override
@@ -39,10 +37,9 @@ public class LoginFirstChecker extends AbstractDateBasedChecker {
 
   @Override
   protected void doCheckAndPublish(Member member, Challenge challenge, LocalDate targetDate) {
-    // 이 사람이 잘 로그인했는가?
+    // 이 사람이 오늘 잘 로그인했으면, MCD 추가
     if (member.getLastLoginDate().toLocalDate().equals(LocalDate.now())) {
-      // MCD 생성 이벤트 발행
-      publishMCDCreateEvent(member, challenge, targetDate);
+      publishMCDCreateEvent(member, challenge,  targetDate);
     }
   }
 }
