+++ everyday/src/main/java/com/nyam/everyday/module/challenge/checker/checker_regular/stamp/FirstStampChecker.java
@@ -1,4 +1,4 @@
-package com.nyam.everyday.module.challenge.checker.regular_checker.meal_log;
+package com.nyam.everyday.module.challenge.checker.checker_regular.stamp;
 
 import com.nyam.everyday.module.challenge.checker.AbstractCountBasedChecker;
 import com.nyam.everyday.module.challenge.checker.service.ChallengeCheckService;
@@ -6,39 +6,39 @@ import com.nyam.everyday.module.challenge.entity.Challenge;
 import com.nyam.everyday.module.challenge.entity.ChallengeCode;
 import com.nyam.everyday.module.challenge.entity.ChallengeTag;
 import com.nyam.everyday.module.challenge.repository.ChallengeRepository;
-import com.nyam.everyday.module.meal.repository.MealLogRepository;
 import com.nyam.everyday.module.member.entity.Member;
+import com.nyam.everyday.module.mission.repository.DailyMissionStampRepository;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 
 @Component
-public class MealLogFirstChecker extends AbstractCountBasedChecker {
+public class FirstStampChecker extends AbstractCountBasedChecker {
 
-  private final MealLogRepository mealLogRepository;
+  private final DailyMissionStampRepository dailyMissionStampRepository;
 
-  protected MealLogFirstChecker(
-          ChallengeRepository challengeRepository,
-          ChallengeCheckService challengeCheckService,
-          ApplicationEventPublisher publisher,
-          MealLogRepository mealLogRepository
+  protected FirstStampChecker(
+      ChallengeRepository challengeRepository,
+      ChallengeCheckService challengeCheckService,
+      ApplicationEventPublisher publisher,
+      DailyMissionStampRepository dailyMissionStampRepository
   ) {
     super(challengeRepository, challengeCheckService, publisher);
-    this.mealLogRepository = mealLogRepository;
+    this.dailyMissionStampRepository = dailyMissionStampRepository;
   }
 
   @Override
   public ChallengeCode getChallengeCode() {
-    return ChallengeCode.MEAL_LOG_FIRST;
+    return ChallengeCode.STAMP_FIRST;
   }
 
   @Override
   public ChallengeTag getChallengeTag() {
-    return ChallengeTag.MEAL_LOG;
+    return ChallengeTag.STAMP;
   }
 
   @Override
   public long getProgress(Member member, Challenge challenge) {
-    // 식단 기록 갯수 세서 반환
-    return mealLogRepository.getCountByMemberId(member.getMemberId());
+    // 도장 횟수를 반환한다.
+    return dailyMissionStampRepository.getCountByMemberId(member.getMemberId());
   }
 }
